crypto/rand module code refactoring
